+++ yggdrasil_decision_forests/utils/model_analysis.cc
@@ -714,6 +714,11 @@ absl::StatusOr<proto::AnalysisResult> Analyse(
     return absl::InvalidArgumentError("The dataset is empty.");
   }
 
+  const absl::optional<float> maximum_duration_seconds =
+      options.has_maximum_duration_seconds()
+          ? options.maximum_duration_seconds()
+          : absl::optional<float>{};
+
   // Try to create a fast engine.
   const model::AbstractModel* effective_model = &model;
   auto engine_or = model.BuildFastEngine();
@@ -737,11 +742,12 @@ absl::StatusOr<proto::AnalysisResult> Analyse(
                          /*flag_2d=*/false,
                          /*flag_2d_categorical_numerical=*/false));
 
-    ASSIGN_OR_RETURN(*analysis.mutable_pdp_set(),
-                     utils::ComputePartialDependencePlotSet(
-                         dataset, *effective_model, attribute_idxs,
-                         options.pdp().num_numerical_bins(),
-                         options.pdp().example_sampling()));
+    ASSIGN_OR_RETURN(
+        *analysis.mutable_pdp_set(),
+        utils::ComputePartialDependencePlotSet(
+            dataset, *effective_model, attribute_idxs,
+            options.pdp().num_numerical_bins(),
+            options.pdp().example_sampling(), maximum_duration_seconds));
   }
 
   // Conditional Expectation Plot
@@ -752,11 +758,12 @@ absl::StatusOr<proto::AnalysisResult> Analyse(
                          /*flag_2d=*/false,
                          /*flag_2d_categorical_numerical=*/false));
 
-    ASSIGN_OR_RETURN(*analysis.mutable_cep_set(),
-                     utils::ComputeConditionalExpectationPlotSet(
-                         dataset, *effective_model, attribute_idxs,
-                         options.cep().num_numerical_bins(),
-                         options.cep().example_sampling()));
+    ASSIGN_OR_RETURN(
+        *analysis.mutable_cep_set(),
+        utils::ComputeConditionalExpectationPlotSet(
+            dataset, *effective_model, attribute_idxs,
+            options.cep().num_numerical_bins(),
+            options.cep().example_sampling(), maximum_duration_seconds));
   }
 
   // TODO: Implement permuted variable importances for anomaly detection.